+++ cpp/CMakeLists.txt
@@ -262,6 +262,21 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libmps_parser)
 set(CMAKE_LIBRARY_PATH ${CMAKE_CURRENT_BINARY_DIR}/libmps_parser/)
 
 
+execute_process(
+  COMMAND git rev-parse --short HEAD
+  WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
+  OUTPUT_VARIABLE GIT_COMMIT_HASH
+  OUTPUT_STRIP_TRAILING_WHITESPACE
+)
+message("-- Building with GIT_COMMIT_HASH = '${GIT_COMMIT_HASH}'")
+
+
+list(JOIN CMAKE_CUDA_ARCHITECTURES "," JOINED_CUDA_ARCHITECTURES) # ';' breaks compile_definitions, replace it
+target_compile_definitions(cuopt PUBLIC
+  CUOPT_GIT_COMMIT_HASH="${GIT_COMMIT_HASH}"
+  CUOPT_CUDA_ARCHITECTURES="${JOINED_CUDA_ARCHITECTURES}"
+  CUOPT_CPU_ARCHITECTURE="${CMAKE_SYSTEM_PROCESSOR}")
+
 target_link_libraries(cuopt
   PUBLIC
   CUDA::cublas
@@ -379,20 +394,6 @@ if(NOT BUILD_LP_ONLY)
     "$<$<COMPILE_LANGUAGE:CUDA>:${CUOPT_CUDA_FLAGS}>"
   )
 
-execute_process(
-    COMMAND git rev-parse --short HEAD
-    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
-    OUTPUT_VARIABLE GIT_COMMIT_HASH
-    OUTPUT_STRIP_TRAILING_WHITESPACE
-)
-message("-- Building with GIT_COMMIT_HASH = '${GIT_COMMIT_HASH}'")
-
-list(JOIN CMAKE_CUDA_ARCHITECTURES "," JOINED_CUDA_ARCHITECTURES) # ';' breaks compile_definitions, replace it
-target_compile_definitions(cuopt_cli PUBLIC
-  CUOPT_GIT_COMMIT_HASH="${GIT_COMMIT_HASH}"
-  CUOPT_CUDA_ARCHITECTURES="${JOINED_CUDA_ARCHITECTURES}"
-  CUOPT_CPU_ARCHITECTURE="${CMAKE_SYSTEM_PROCESSOR}")
-
   target_include_directories(cuopt_cli
     PRIVATE
     "${CMAKE_CURRENT_SOURCE_DIR}/src"
@@ -426,10 +427,6 @@ if(BUILD_MIP_BENCHMARKS AND NOT BUILD_LP_ONLY)
     PRIVATE "$<$<COMPILE_LANGUAGE:CXX>:${CUOPT_CXX_FLAGS}>"
     "$<$<COMPILE_LANGUAGE:CUDA>:${CUOPT_CUDA_FLAGS}>"
   )
-  target_compile_definitions(solve_MIP PUBLIC
-    "CUOPT_GIT_COMMIT_HASH=\"${GIT_COMMIT_HASH}\""
-    "CUOPT_CUDA_ARCHITECTURES=\"${JOINED_CUDA_ARCHITECTURES}\""
-    "CUOPT_CPU_ARCHITECTURE=\"${CMAKE_SYSTEM_PROCESSOR}\"")
   target_link_libraries(solve_MIP
     PUBLIC
     cuopt
@@ -446,10 +443,6 @@ if(BUILD_LP_BENCHMARKS)
     PRIVATE "$<$<COMPILE_LANGUAGE:CXX>:${CUOPT_CXX_FLAGS}>"
     "$<$<COMPILE_LANGUAGE:CUDA>:${CUOPT_CUDA_FLAGS}>"
   )
-  target_compile_definitions(solve_LP PUBLIC
-    "CUOPT_GIT_COMMIT_HASH=\"${GIT_COMMIT_HASH}\""
-    "CUOPT_CUDA_ARCHITECTURES=\"${JOINED_CUDA_ARCHITECTURES}\""
-    "CUOPT_CPU_ARCHITECTURE=\"${CMAKE_SYSTEM_PROCESSOR}\"")
   target_link_libraries(solve_LP
     PUBLIC
     cuopt
