+++ cpp/src/CMakeLists.txt
@@ -14,7 +14,8 @@
 # limitations under the License.
 
 set(UTIL_SRC_FILES ${CMAKE_CURRENT_SOURCE_DIR}/utilities/seed_generator.cu
-                   ${CMAKE_CURRENT_SOURCE_DIR}/utilities/logger_helper.cpp)
+                   ${CMAKE_CURRENT_SOURCE_DIR}/utilities/logger_helper.cpp
+                   ${CMAKE_CURRENT_SOURCE_DIR}/utilities/version_info.cpp)
 
 add_subdirectory(linear_programming)
 add_subdirectory(math_optimization)
